# static/swagger.yaml
swagger: '2.0'
info:
  description: >-
    This API is a set of functions that provides consumer and supplier with all
    ecosystem services
  version: 1.0.4
  title: Internet of Things Analytical Ecosystem API
  contact: {}
host: 'localhost:5000'
basePath: /
...
schemes:
  - http
securityDefinitions:
  GoogleIdToken:
    description: >-
      [Google Identity Platform](https://developers.google.com/identity/)
      authentication with id_token as authentication key. The security
      definision in OpenAPI 2.0 lacks the 'Bearer ' string before <id_token> as
      described in [RFC 6750 section
      2.1](https://tools.ietf.org/html/rfc6750#section-2.1) Please insert
      'Bearer ' manualy. Remember to insert required header in your application
      requests 'Authentication &#58; Bearer <id_token>'
    type: apiKey
    in: header
    name: Authorization
paths:
  /bill/:
    get:
      tags:
        - Billing
        - Consumer API
        - Supplier API
      summary: Receive a billing report
      description: Receive billing report
      operationId: bill_rcv
      produces:
        - application/json
      security:
        - GoogleIdToken: []
      parameters:
        - in: query
          name: begin
          type: string
          description: Begin of the time span of a report (yyyymmdd)
          required: false
          x-example: '20180410'
        - in: query
          name: end
          type: string
          description: End of a time span of a report (yyyymmdd)
          required: false
          x-example: '20180410'
      responses:
        '200':
          description: Billing Report
          schema:
            $ref: '#/definitions/BillingReport'
  '/bill/result/{resultsetId}':
    get:
      tags:
        - Billing
        - Consumer API
      summary: Receive a billing summary for result set
      description: Receive a billing summary for result set
      operationId: bill_rcv(resultsetId)
      produces:
        - application/json
      security:
        - GoogleIdToken: []
      parameters:
        - name: resultsetId
          in: path
          required: true
          type: string
        - in: query
          name: begin
          type: string
          description: Begin of the time span of a report (yyyymmdd)
          required: false
          x-example: '20180410'
        - in: query
          name: end
          type: string
          description: End of a time span of a report (yyyymmdd)
          required: false
          x-example: '20180410'
      responses:
        '200':
          description: Billing result set summary
          schema:
            $ref: '#/definitions/BillingReport'
  '/bill/algorithm/{algorithmId}':
    get:
      tags:
        - Billing
        - Supplier API
      summary: Receive a billing summary for algorith
      description: Receive a billing summary for algorithm
      operationId: bill_rcv(algorithmId)
      produces:
        - application/json
      security:
        - GoogleIdToken: []
      parameters:
        - name: algorithmId
          in: path
          required: true
          type: string
        - in: query
          name: begin
          type: string
          description: Begin of the time span of a report (yyyymmdd)
          required: false
          x-example: '20180410'
        - in: query
          name: end
          type: string
          description: End of a time span of a report (yyyymmdd)
          required: false
          x-example: '20180410'
      responses:
        '200':
          description: Billing algorithm summary
          schema:
            $ref: '#/definitions/BillingReport'
  /user/:
    get:
      tags:
        - User
        - Admin
        - Consumer API
        - Supplier API
      summary: Get user information
      description: Get user info - requres authorization
      operationId: get_user
      produces:
        - application/json
      security:
        - GoogleIdToken: []
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/User'
        '400':
          description: Invalid username
        '404':
          description: User not found
    delete:
      tags:
        - User
        - Admin
      summary: Delete user
      description: This can only be done by the Administrator.
      operationId: deleteUser
      produces:
        - application/json
      security:
        - GoogleIdToken: []
      responses:
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
    post:
      tags:
        - User
        - Admin
      summary: Create a new user
      description: This can only be done by admin.
      operationId: create_user
      produces:
        - application/json
      security:
        - GoogleIdToken: []
      responses:
        '200':
          description: OK
  /datasets:
    get:
      tags:
        - Consumer API
        - Dataset
      summary: Returns information about the stored Datasets
      description: >-
        The Datasets endpoint returns information about the stored Datasets. The
        response includes the display name and other details about each dataset.
        It also allows full-text search of tags.
      operationId: getDatasets
      consumes: []
      produces:
        - application/json
      security:
        - GoogleIdToken: []
      parameters:
        - name: tags
          in: query
          description: Comma separated Tags for search query.
          required: false
          type: string
          x-example: 'sum two items, sum multiple items'
      responses:
        '200':
          description: An array of datasets
          schema:
            type: array
            items:
              $ref: '#/definitions/Datasets'
        '401':
          description: No authentication. Need to login.
        '403':
          description: Not authorized to post new datasets
    post:
      tags:
        - Dataset
        - Consumer API
      summary: Add a Dataset
      description: Add a Dataset
      operationId: postDataset
      consumes:
        - application/json
      produces:
        - application/json
      security:
        - GoogleIdToken: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/DatasetBody'
      responses:
        '200':
          description: OK - returns datasetID
          schema:
            $ref: '#/definitions/Datasets'
        '401':
          description: No authentication. Need to login.
        '403':
          description: Not authorized to post new datasets
  '/datasets/{datasetID}':
    get:
      tags:
        - Dataset
        - Consumer API
      summary: Returns a given dataset
      description: Returns a given dataset
      operationId: getDatasetsDataset
      consumes: []
      produces:
        - application/json
      security:
        - GoogleIdToken: []
      parameters:
        - name: datasetID
          in: path
          description: ID of a dattaset
          required: true
          type: string
      responses:
        '200':
          description: Data of the dataset
          schema:
            $ref: '#/definitions/Dataset'
        '401':
          description: No authentication. Need to login.
        '403':
          description: Not authorized to read that datasets
        '404':
          description: The given datasetID was not found in collection
    delete:
      tags:
        - Dataset
        - Consumer API
      summary: Deletes a given dataset
      description: Deletes a given dataset
      operationId: deleteDatasetsDataset
      consumes: []
      produces:
        - application/json
      security:
        - GoogleIdToken: []
      parameters:
        - name: datasetID
          in: path
          description: ID of a dattaset
          required: true
          type: string
      responses:
        '200':
          description: Deletion completed
        '401':
          description: No authentication. Need to login.
        '403':
          description: Not authorized to delete that dataset
        '404':
          description: The given datasetID was not found in collection
    put:
      tags:
        - Dataset
        - Consumer API
      summary: Add etities to dataset
      description: Add etities to dataset
      operationId: addDatasetsData
      consumes:
        - application/json
      produces: []
      security:
        - GoogleIdToken: []
      parameters:
        - name: datasetID
          in: path
          description: ID of a dataset
          required: true
          type: string
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Dataset'
      responses:
        '200':
          description: OK
        '400':
          description: Malformed data
          schema:
            $ref: '#/definitions/Error'
        '401':
          description: No authentication. Need to login.
        '403':
          description: >-
            Not authorized to post data at all or not an owner of a given
            dataset
        '404':
          description: algorithmID not found
  /algorithms/:
    get:
      tags:
        - Consumer API
        - Algorithms
      summary: All available Algorithms descriptions.
      description: >-
        The Algorithms endpoint returns information about the available
        algorithms. The response includes the display name and other details
        about each algorithm. It also allows full-text search of tags.
      operationId: getAlgorithms
      consumes: []
      produces:
        - application/json
      parameters:
        - name: tags
          in: query
          description: >-
            Comma separated Tags for search query No whitespace characters
            allowed in searched tags. No wildcard or substring searches.
          required: false
          type: string
          x-example: 'algorithm1, algorithm2'
      responses:
        '200':
          description: An array of algorithms
          schema:
            type: array
            items:
              $ref: '#/definitions/Algorithms'
    post:
      tags:
        - Algorithms
        - Supplier API
      summary: Add a new Algorithm
      description: Add a new Algorithm
      operationId: postAlgorithms
      consumes:
        - application/json
      produces: []
      security:
        - GoogleIdToken: []
      parameters:
        - in: body
          name: body
          required: false
          schema:
            $ref: '#/definitions/AlgorithmBody'
      responses:
        '200':
          description: OK
        '400':
          description: Malformed data
          schema:
            $ref: '#/definitions/Error'
        '401':
          description: No authentication. Need to login.
        '403':
          description: Not authorized to post new algorithms
        '409':
          description: AlgorithmID already exists.
          schema:
            $ref: '#/definitions/Error'
...
definitions:
...
  Datasets:
    type: object
    properties:
      datasetId:
        type: string
        description: Unique identifier representing a specific dataset.
      datasetSummary:
        type: string
        description: Users description of a dataset
      displayName:
        type: string
        description: Display name of the dataset
      linkURL:
        type: string
        description: detailed url of the dataset.
...
  Error:
    type: object
    properties:
      code:
        type: integer
        format: int32
      fields:
        type: string
      message:
        type: string
...
  Dataset:
    type: object
    properties:
      dataset:
        type: string
        description: Data entity to be added to dataset.
...
  User:
    type: object
    properties:
      userID:
        type: string
        description: >-
          ID from sub field extracted from id_token as described in [Google
          Identity Platform](https://developers.google.com/identity/)
      firstName:
        type: string
      lastName:
        type: string
      email:
        type: string
      phone:
        type: string
      userStatus:
        type: integer
        format: int32
        description: User Status
...
  BillingReport:
    type: object
    description: Report containng billing information
    properties:
      reportData:
        type: string
        description: Complete data of the billing report in xml